pyproject.toml
[project]
name = "forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "einops>=0.8.1",
    "fair-esm>=2.0.0",
    "hydra-core>=1.3.2",
    "lmdb>=1.7.3",
    "pytest>=8.4.1",
    "pytorch-lightning>=2.5.2",
    "raygun",
    "ruff>=0.12.7",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.55.0",
    "wandb>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
raygun = { git = "https://github.com/rohitsinghlab/raygun.git" }

[project.scripts]
forge = "forge.inference.cli:main"
